+++ Changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 -   Build with github actions
+-   Update dependencies
+-   Print execution time for each test in summary and order entries by execution time
 
 ## [1.4.0] 2020-04-15